add zoo like a separate package
